Refactors folders
