+++ src/App.js
@@ -2,7 +2,7 @@ import React from "react";
 import "./assets/styles.scss";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Home from "./views/Home";
-import Details from "./components/Details";
+import Details from "./views/Details";
 function App() {
   return (
     <main>
