+++ src/assets/scss/details.scss
@@ -13,7 +13,6 @@ $button-hover-color: #34b3d3;
 //     top: 0;
 //     display: none;
 //   }
-  
 //   .imageBox:hover .hoverImg {
 //     display: block;
 //   }
