+++ src/assets/styles.scss
@@ -1,6 +1,6 @@
 @import url("https://fonts.googleapis.com/css2?family=Quicksand:wght@300;400;700&display=swap");
 @import "./scss/globals.scss";
-@import "../components/assets/details.scss";
+@import "./scss/details.scss";
 /*
 Cuando vayais a hacer el estilo de un componente solo teneis que enlazarlo aqui así y te olvidas:
