+++ src/views/Details.js
@@ -8,7 +8,7 @@ import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 import 'swiper/css/scrollbar';
 
-import useFindCurrentPet from "./hooks/useCurrentPet";
+import useFindCurrentPet from "../hooks/useCurrentPet";
 
 export default function Details() {
     const { id } = useParams();
